Deduplicate historical Java version tag and upgrade client library
The new way of extracting the Java version doesn't include the internal
build number, so remove that from historical data.

First, upgrade the InfluxDB client library.

This fixes two issues that I reported long ago with Nginx reverse proxy
and basic authentication, so remove the workarounds.

It also has a new feature to talk to the server in a binary protocal
(Message Pack). Apart from being more efficient (which doesn't really
matter much to us) it also doesn't lose track of int vs float
data types when doing a bulk query (like we do in the migrator)
which avoids needing to manually narrow "sampleCount" to a integral
type.